Don't duplicate columns in 'SELECT * ...' queries
Don't duplicate columns in 'SELECT * ...' queries. In queries that
select all table columns, a new AST node will be added for each column
in the result. This behaviour is incorrect, as nested members will be
added to the AST twice, once as part of the orignal object and once as a
stand-alone node. Only the top level AST nodes shoud be added to the
result AST.

	- Update row_callback(...) to detect 'SELECT * ...' queries and
	  only append top-level nodes to result AST.
	- Add regression test