Update result handling to support subqueries
Update result handling to support subqueries. Supporting subqueries
necessitates that we track result rows in a slightly smarter way.  After
a call to rows.Next() our result is not necessarily the last set of
columns that SQLite asked us for in Columns(...). Instead, we have to
rely on the values returned by rows.Scan(...).

At the moment we handle this quite crudely so we can still use the
pretty printer to write our results. This needs improvement.

	- Remove notion of result tracking from jsonTable
	- Update SchemasFromStmt to return both the CREATE TABLE
	  statements and columns in the statement
	- Update Column(...) method to crudely handle alised tables
	- Update main rows.Next() result loop to use rows.Scan(...)
	- Add test case with subquery and uncomment existing JOIN test
	  case